Primera versión funcional

+++ PDFU/PDFU_extract.py
@@ -3,46 +3,58 @@ from pdfrw import PdfReader, PdfWriter
 from pdfrw.findobjs import wrap_object , find_objects
 from pdfrw.objects import PdfName
 from pikepdf import Pdf
-import traceback
-import tempfile
 
-def clean_fun(pdf_path):
-    prepdf=Pdf.open(pdf_path)
-    prepdf.save(pdf_path[:-4]+"_inter.pdf")
-    prepdf.close()
-
-    pdf = PdfReader(pdf_path[:-4]+"_inter.pdf")
-    xobjs=list(find_objects(pdf.pages,valid_subtypes=(PdfName.Form, PdfName.Ima)))
-    páginas=[]
-    for item in xobjs:
-        páginas.append(wrap_object(item, 1000, 0.5*72))
-    if len(xobjs)==0:
-        os.remove(pdf_path[:-4]+"_inter.pdf")
-        output=False
-        return output
-    del xobjs
-    del pdf
-
-    output=pdf_path[:-4]+"_limpio.pdf"
-    writer = PdfWriter(output)
-    writer.addpages(páginas)
-    del páginas
-    writer.write()
-    del writer
-
-    os.remove(pdf_path[:-4]+"_inter.pdf")
-
-    return output
-
-
-def clean():
-    root = "C:/users/yomis/python/gulag_cleaner_standalone/"
-    file="test.pdf"
+def deembed(pdf_path):
+    '''
+    Deembeds the pdf and creates a new PDF in the same folder with each embedded page
+    in a new page.
+    
+    Args:
+        pdf_path: The path where the pdf file is located.
+        
+    Returns:
+        return_msg: Dict. with three values:
+            Success: bool indicating whether the process was successful.
+            return_path: If successful, returns the path of the deembedded file.
+            Error: If unsuccessful, returns a description of the error.
+    '''
+    
+    return_msg={"Success":False,"return_path":"","Error":""}
     try:
-        outputpath=clean_fun(root+file)
-        print(outputpath)
-    except Exception as e:
-        print(e)
-        traceback.print_exc()
+        if pdf_path[:-4]!=".pdf":
+            return_msg["Success"]=False
+            return_msg["Error"]="File is not a .pdf file."
+            return return_msg
+        prepdf=Pdf.open(pdf_path)
+        prepdf.save(pdf_path[:-4]+"_inter.pdf")
+        prepdf.close()
+
+        pdf = PdfReader(pdf_path[:-4]+"_inter.pdf")
+        xobjs=list(find_objects(pdf.pages,valid_subtypes=(PdfName.Form, PdfName.Dummy)))
+        páginas=[]
+        for item in xobjs:
+            páginas.append(wrap_object(item, 1000, 0.5*72))
+        if len(xobjs)==0:
+            os.remove(pdf_path[:-4]+"_inter.pdf")
+            return_msg["Success"]=False
+            return_msg["Error"]="No embedded pages found."
+            return return_msg
+
+        output=pdf_path[:-4]+"_deembedded.pdf"
+        writer = PdfWriter(output)
+        writer.addpages(páginas)
+        writer.write()
+
+        os.remove(pdf_path[:-4]+"_inter.pdf")
+        return_msg["Success"]=True
+        return_msg["return_path"]=output
+        return return_msg
+    except:
+        return_msg["Success"]=False
+        return_msg["Error"]="Unknown internal error."
+        return return_msg
+
+
 if __name__ == "__main__":
-    clean()
+    print('Call from the "pdfu" command.')
+
